pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cybertco"
version = "0.1.0"
description = "Cybersecurity program costing & forecasting (CAPEX/OPEX) with reporting"
readme = "README.md"
authors = [{name = "Moazzam (Mo)"}, {name="Contributors"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.0",
  "pyyaml>=6.0",
  "matplotlib>=3.8"
]

[project.scripts]
cybertco = "cybertco.cli:main"
